Move unit system enums to corecel
