+++ src/celeritas/field/RZMapFieldInputIO.json.cc
@@ -12,6 +12,7 @@
 #include <string>
 #include <vector>
 
+#include "corecel/Types.hh"
 #include "corecel/cont/Range.hh"
 #include "corecel/io/JsonUtils.json.hh"
 #include "corecel/io/Logger.hh"
